Improving hiding of Shorts
The styling is prepended to the html element instead of being appended.
This seemed to make an improvement in speed but I am unsure if it
actually makes a difference.

Using display: none!important; instead of changing both visibility,
width and height results in the hidden elements not being a part of the
rendered DOM. This solves the issue of Shorts in subscriptions still
taking up place. It also simplifies the CSS for hiding elements.