Hide suggested posts on instagram
First, suggested posts on the home screen are hidden. This includes the
suggested posts themselves, the text saying "Suggested posts" and the
loading logo when more posts are loaded.

Second, links to more posts under a post/reel are hidden. This is done
in order to prevent an infinite clicking to new posts/reels. When
viewing a post/reel over the Explore page, Search page etc., it is
still possible to use the arrows to go to the previous and next post/
reel.