Remove suggested posts in Facebook feed
Using MutationObserver to observe changes in the home feed DOM and
remove nodes where the first three words of the innerText are
"Suggested for you".

The MutationObserver is initialized after the window (and therefore the
DOM) has finished loading. This is done so the home feed node can be
accessed. An additional check is used in order to remove "Suggested for
you" posts that are already loaded.